pyproject.toml
[tool.poetry]
name = "github-meme-bot"
version = "0.0.1"
description = "Make meme in github comments"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MeetWq/github-meme-bot"
repository = "https://github.com/MeetWq/github-meme-bot"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.3.0"
nonebot-adapter-github = "^0.4.1"
meme-generator = ">=0.0.15,<0.1.0"
httpx = ">=0.20.0,<1.0.0"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "E501",   # line-too-long
  "UP037",  # quoted-annotation
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "basic"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
